+++ mobile_wallet/android/mobile_wallet_lib/src/main/java/com/concordium/mobile_wallet_lib/wallet.kt
@@ -19,6 +19,8 @@ external fun check_account_address(input: String) : Boolean
 external fun link_check(input: String) : String
 external fun get_identity_keys_and_randomness(input: String) : ReturnValue
 external fun get_account_keys_and_randomness(input: String) : ReturnValue
+external fun sign_transaction(input: String) : ReturnValue
+external fun transaction_to_json(input: String) : ReturnValue
 
 fun loadWalletLib() {
     System.loadLibrary("mobile_wallet")